# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CampaignCell API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: segment
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignListResponse"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /v1/campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete campaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
...
  /v1/campaigns/{id}/publish:
    post:
      operationId: publishCampaign
      tags: [campaigns]
      summary: Publish a campaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /v1/campaigns/{id}/archive:
    post:
      operationId: archiveCampaign
      tags: [campaigns]
      summary: Archive a campaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  # ─── OPTIMIZATION CASES ─────────────────────────────────────────────────────
...
  /v1/subscribers/{id}/campaigns:
    get:
      operationId: getSubscriberCampaigns
      tags: [subscribers]
      summary: Get campaigns for a subscriber (personalized offers)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Personalized campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriberCampaignListResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # ── AUTH SCHEMAS ──────────────────────────────────────────────────────────
...
    Campaign:
      type: object
      required: [id, name, type, status, segment, priority, discount, startDate, endDate, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [EK_PAKET, TARIFE_YUKSELTME, CIHAZ_FIRSATI, SADAKAT]
        status:
          type: string
          enum: [DRAFT, ACTIVE, MANUAL_OPTIMIZATION_REQUIRED, OPTIMIZING, AB_TESTING, OPTIMIZED, PUBLISHED, ARCHIVED]
        segment:
          type: string
          enum: [YUKSEK_DEGER, RISKLI_KAYIP, YENI_ABONE, PASIF, BELIRSIZ]
        priority:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        discount:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        recommendationScore:
          type: ["number", "null"]
        conversionProbability:
          type: ["number", "null"]
        aiReasoning:
          type: ["string", "null"]
        isAiAnalyzed:
          type: boolean
        conversionRate:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CampaignInput:
      type: object
      required: [name, type, discount, startDate, endDate]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [EK_PAKET, TARIFE_YUKSELTME, CIHAZ_FIRSATI, SADAKAT]
        segment:
          type: string
          enum: [YUKSEK_DEGER, RISKLI_KAYIP, YENI_ABONE, PASIF, BELIRSIZ]
        discount:
          type: number
        startDate:
          type: string
        endDate:
          type: string
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        segment:
          type: string
          enum: [YUKSEK_DEGER, RISKLI_KAYIP, YENI_ABONE, PASIF, BELIRSIZ]
        priority:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        discount:
          type: number
    CampaignListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Campaign"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    SubscriberCampaign:
      type: object
      required: [id, campaignId, name, type, discount, recommendationScore, conversionProbability, aiReasoning, status]
      properties:
        id:
          type: string
        campaignId:
          type: string
        name:
          type: string
        type:
          type: string
        discount:
          type: number
        validUntil:
          type: ["string", "null"]
        recommendationScore:
          type: number
        conversionProbability:
          type: number
        aiReasoning:
          type: string
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, RATED]
        rating:
          type: ["integer", "null"]
        rejectionReason:
          type: ["string", "null"]
...
    SubscriberCampaignListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SubscriberCampaign"
...